fix: correct data types for station and line name lengths and adjust record reading logic

+++ project/domain/subway_line_record.h
@@ -11,9 +11,9 @@ struct SubwayLineRecord {
     uint32_t destinationDistant;
     uint32_t interactionStationID;
     uint32_t interactionLineID;
-    size_t stationNameLength;
+    uint32_t stationNameLength;
     char *stationName;
-    size_t lineNameLength;
+    uint32_t lineNameLength;
     char *lineName;
 };
